폰트 업로드

+++ Assets/Scripts/Player/PlayerWeapons.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerWeapons : MonoBehaviour
 {
     [Tooltip("데미지")]
-    [SerializeField] internal int damage;
+    [SerializeField] internal float damage;
     [Tooltip("공격 시간")]
     [SerializeField] internal float time;
     [SerializeField] internal float saveDamage;
